requirements.txt
 Django==4.2.5
 djangorestframework==3.14.0
 djangorestframework-simplejwt==5.3.0
 psycopg2-binary==2.9.7
 celery[redis]==5.3
 flower==2.0.1
 celery_singleton==0.3.1
 drf-spectacular==0.26.4
requests==2.31.0